Creacion de menuUser, correcion de App.js y agregacion de token de verificacion de login para acceder a ciertas paginas
