+++ mindpeace-frontend/src/components/PrivateRoute.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Navigate } from 'react-router-dom';
+
+function PrivateRoute({ children }) {
+    const token = localStorage.getItem('authToken');
+
+    if (!token) {
+        return <Navigate to="/login" />;
+    }
+
+    return children;
+}
+
+
+export default PrivateRoute;
